chore: split StoryGroupTree component
